+++ src/Firebase/Http/Responses.php
@@ -26,7 +26,7 @@ final class Responses implements IteratorAggregate
      *
      * @return Traversable<ResponseInterface>|ResponseInterface[]
      */
-    public function getIterator(): iterable
+    public function getIterator(): Traversable
     {
         yield from $this->responses;
     }